+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## New Features
 
+## New features
+
+- `add_definition()` (#178), `update_definition()` (#127), and `delete_definition()` functions for FeatureServer and FeatureLayer objects.
+
 ## Bug Fixes
 
 - Improve `update_features()` with an error message when the `objectid` is not an `integer` ([#250](https://github.com/R-ArcGIS/arcgislayers/issues/250))